refactor: Move filter reducer to a separate file
